requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
blis==0.4.1
catalogue==1.0.0
celery==4.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cymem==2.0.3
decorator==4.4.2
Django==2.2.3
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-celerybeat-status==0.0.10
django-cors-headers==3.2.1
django-timezone-field==4.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
jellyfish==0.7.2
joblib==0.14.1
kombu==4.6.8
langdetect==1.0.8
murmurhash==1.0.2
networkx==2.4
nltk==3.5
numpy==1.18.2
oauthlib==3.1.0
plac==1.1.3
preshed==3.0.2
psycopg2==2.8.5
PySocks==1.7.1
python-crontab==2.4.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
segtok==1.5.9
six==1.14.0
sklearn==0.0
spacy>=2.0.0,<3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz#en_core_web_sm
sqlparse==0.3.1
srsly==1.0.2
thinc==7.4.0
tqdm==4.45.0
tweepy==3.8.0
urllib3==1.25.8
vine==1.3.0
wasabi==0.6.0
Wikipedia-API==0.5.4
zipp==3.1.0
pattern
monthdelta
gensim==3.8.2
